# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management System",
        "description": "This is an api documentation for a user management system.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "test@codelab-ums-api.herokuapp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "operationId": "App\\Http\\Controllers\\UserController::index",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "default": {
                        "description": "error"
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Fetch user",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user"
                    },
                    "default": {
                        "description": "error"
                    }
                }
            }
        },
        "api/user/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "App\\Http\\Controllers\\UserController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a user"
                    },
                    "default": {
                        "description": "error"
                    }
                }
            }
        }
    }
}